Moves FindUnitServiceNames to service package and adds unit test.
